apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
boto3
openai
PyJWT[crypto]
pypdf
python-docx

